dynamic-threads: fix crash when disabled user tries to login
git-svn-id: http://streamboard.de.vu/svn/oscam/branches/dynamic-threads@5597 4b0bc96b-bc66-0410-9d44-ebda105a78c1

+++ global-functions.h
@@ -291,7 +291,7 @@ extern void network_tcp_connection_close(struct s_reader *);
 extern void clear_reader_pipe(struct s_reader * reader);
 extern void block_connect(struct s_reader *rdr);
 extern int32_t is_connect_blocked(struct s_reader *rdr);
-void cs_add_entitlement(struct s_reader *rdr, uint16_t caid, uint32_t provid, uint16_t id, uint16_t class, time_t start, time_t end, uint8_t type);
+void cs_add_entitlement(struct s_reader *rdr, uint16_t caid, uint32_t provid, uint64_t id, uint32_t class, time_t start, time_t end, uint8_t type);
 extern void cs_clear_entitlement(struct s_reader *rdr);
 
 /* ===========================
